Changed prices in DB to floats

+++ models/alerts.js
@@ -5,8 +5,8 @@ module.exports = (sequelize, DataTypes) => {
     userID: DataTypes.INTEGER,
     title: DataTypes.STRING,
     gameID: DataTypes.INTEGER,
-    desiredprice: DataTypes.INTEGER,
-    setprice: DataTypes.INTEGER
+    desiredprice: DataTypes.FLOAT,
+    setprice: DataTypes.FLOAT
   }, {})
   alerts.associate = function (models) {
     models.alerts.belongsTo(models.users, {
